Save RS485 start communication
