+++ README.md
@@ -6,5 +6,6 @@
 - [x] Remove existing subscriber
 - [x] Change serial_connection to rs485 connecection
 - [X] add publisher for the mission switch AND kill switch
-- [ ] Look at the SendRS485Message in Sonia_common
+- [x] Look at the SendRS485Message in Sonia_common
+- [ ] Split the different part of the project
 - [ ] Write new message for publisher
