+++ src/RS485Provider.cpp
@@ -10,34 +10,9 @@ namespace rs485_port_manager
     RS485Provider::RS485Provider()
         : Node("rs485_provider"), _rs485Connection("/dev/RS485", B115200, false), _thread_control(true)
     {
-        try
-        {
-            auv = std::getenv("AUV");
-            if (strcmp(auv, "AUV8")==0 || strcmp(auv, "LOCAL")==0)
-            {
-                ESC_SLAVE = _SlaveId::SLAVE_PWR_MANAGEMENT;
-                RCLCPP_INFO(this->get_logger(), "Using AUV8 port");
-            }
-            else if(strcmp(auv, "AUV7")==0)
-            {
-                ESC_SLAVE = _SlaveId::SLAVE_ESC;
-                RCLCPP_INFO(this->get_logger(), "Using AUV7 port");
-            }
-            else 
-            {
-                ESC_SLAVE = _SlaveId::SLAVE_ESC;
-                RCLCPP_INFO(this->get_logger(), "Using default port AUV7");
-            }
 
-        }
-        catch (...)
-        {
-            ESC_SLAVE = _SlaveId::SLAVE_ESC;
-        }
-
-        group1 = this->create_callback_group(rclcpp::CallbackGroupType::MutuallyExclusive);
         auto sub_opt = rclcpp::SubscriptionOptions();
-        sub_opt.callback_group = group1;
+        sub_opt.callback_group = this->create_callback_group(rclcpp::CallbackGroupType::MutuallyExclusive);
 
         _reader = std::thread(std::bind(&RS485Provider::readData, this));
         _writer = std::thread(std::bind(&RS485Provider::writeData, this));
@@ -45,31 +20,13 @@ namespace rs485_port_manager
         _publisherKill = this->create_publisher<sonia_common_ros2::msg::KillStatus>("/provider_rs485/kill_status", 10);
         _publisherMission =
             this->create_publisher<sonia_common_ros2::msg::MissionStatus>("/provider_rs485/mission_status", 10);
-        _publisherMotorVoltages =
-            this->create_publisher<sonia_common_ros2::msg::MotorPowerMessages>("/provider_power/motor_voltages", 10);
-        _publisherMotorCurrents =
-            this->create_publisher<sonia_common_ros2::msg::MotorPowerMessages>("/provider_power/motor_currents", 10);
-        _publisherMotorTemperature = this->create_publisher<sonia_common_ros2::msg::MotorPowerMessages>(
-            "/provider_power/motor_temperatures", 10);
-        _publisherBatteryVoltages = this->create_publisher<sonia_common_ros2::msg::BatteryPowerMessages>(
-            "/provider_power/battery_voltages", 10);
-        _publisherBatteryCurrents = this->create_publisher<sonia_common_ros2::msg::BatteryPowerMessages>(
-            "/provider_power/battery_currents", 10);
-        _publisherBatteryTemperature = this->create_publisher<sonia_common_ros2::msg::BatteryPowerMessages>(
-            "/provider_power/battery_temperatures", 10);
-        _publisherMotorFeedback =
-            this->create_publisher<sonia_common_ros2::msg::MotorFeedback>("/provider_power/motor_feedback", 10);
         _actuatorService = this->create_service<sonia_common_ros2::srv::ActuatorService>(
             "/provider_actuator/do_action", std::bind(&RS485Provider::processActuatorRequest, this, _1, _2));
         _timerKillMission = this->create_wall_timer(500ms, std::bind(&RS485Provider::pollKillMission, this));
         // _timerPowerRequest= this->create_wall_timer(500ms, std::bind(&RS485Provider::pollPower, this));
-
-        _publisherThrusterPwm =
-            this->create_publisher<sonia_common_ros2::msg::MotorPwm>("/provider_thruster/thruster_pwm", 10);
-        _subscriberThrusterPwm = this->create_subscription<sonia_common_ros2::msg::MotorPwm>(
-            "/provider_thruster/thruster_pwm", 10, std::bind(&RS485Provider::PwmCallback, this, _1), sub_opt);
-        _subscriberMotorOnOff = this->create_subscription<std_msgs::msg::Bool>(
-            "/provider_power/activate_motors", 10, std::bind(&RS485Provider::EnableDisableMotors, this, _1), sub_opt);
+        
+        _publisherRS485 = this->create_subscription<sonia_common_ros2::msg::RS485msg>(
+            "/rs485/msgToSend", 10, std::bind(&RS485Provider::RS485callback, this, _1), sub_opt);
     }
 
     // node destructor
@@ -83,6 +40,24 @@ namespace rs485_port_manager
             _rs485Connection.Flush();
         }
         return res;
+    }
+
+    void RS485Provider::RS485callback(const sonia_common_ros2::msg::RS485msg &msg)
+    {
+        queueObject ser;
+        ser.cmd = msg.cmd;
+        ser.slave = msg.slave;
+        ser.data.clear();
+
+        std::vector<uint8_t> uint8Vector;
+        uint8Vector.resize(msg.data.size());
+        
+        std::transform(msg.data.begin(), msg.data.end(), uint8Vector.begin(),
+                   [](char c) { return static_cast<uint8_t>(c); });
+        ser.data = uint8Vector;
+        
+        _writerQueue.push_back(ser);
+        _cvReaderWriter.notify_all();
     }
 
     void RS485Provider::pollKillMission()
@@ -174,227 +149,6 @@ namespace rs485_port_manager
         _publisherMission->publish(state);
     }
 
-    void RS485Provider::publishMotor(uint8_t cmd, std::vector<float> data)
-    {
-        /*if (data.size() != 8)
-        {
-            return;
-        }*/
-        sonia_common_ros2::msg::MotorPowerMessages msg;
-        msg.motor1 = data[0];
-        msg.motor2 = data[1];
-        msg.motor3 = data[2];
-        msg.motor4 = data[3];
-        msg.motor5 = data[4];
-        msg.motor6 = data[5];
-        msg.motor7 = data[6];
-        msg.motor8 = data[7];
-
-        switch (cmd)
-        {
-            case _Cmd::CMD_VOLTAGE:
-                _publisherMotorVoltages->publish(msg);
-                break;
-            case _Cmd::CMD_CURRENT:
-                _publisherMotorCurrents->publish(msg);
-                break;
-            case _Cmd::CMD_TEMPERATURE:
-                _publisherMotorTemperature->publish(msg);
-                break;
-            default:
-                break;
-        }
-    }
-
-    void RS485Provider::publishBattery(uint8_t cmd, float *data)
-    {
-        sonia_common_ros2::msg::BatteryPowerMessages msg;
-        msg.battery1 = data[0];
-        msg.battery2 = data[1];
-
-        switch (cmd)
-        {
-            case _Cmd::CMD_VOLTAGE:
-                _publisherBatteryVoltages->publish(msg);
-                break;
-            case _Cmd::CMD_CURRENT:
-                _publisherBatteryCurrents->publish(msg);
-                break;
-            case _Cmd::CMD_TEMPERATURE:
-                _publisherBatteryTemperature->publish(msg);
-                break;
-            default:
-                break;
-        }
-    }
-
-    void RS485Provider::publishMotorFeedback(std::vector<uint8_t> data)
-    {
-        sonia_common_ros2::msg::MotorFeedback msg;
-        msg.motor1 = data[0];
-        msg.motor2 = data[1];
-        msg.motor3 = data[2];
-        msg.motor4 = data[3];
-        msg.motor5 = data[4];
-        msg.motor6 = data[5];
-        msg.motor7 = data[6];
-        msg.motor8 = data[7];
-        _publisherMotorFeedback->publish(msg);
-    }
-
-    void RS485Provider::processPowerManagement(const uint8_t cmd, const std::vector<uint8_t> data)
-    {
-        std::vector<float> motorData;
-        motorData.reserve(10);
-        float batteryData[2];
-
-        switch (cmd)
-        {
-            case _Cmd::CMD_VOLTAGE:
-
-                if (convertBytesToFloat(data, motorData, nb_thruster + nb_battery) < 0)
-                {
-                    std::cerr << "ERROR in the message. Dropping VOLTAGE packet" << std::endl;
-                    return;
-                }
-
-                batteryData[0] = motorData[motorData.size() - 2];
-                batteryData[1] = motorData[motorData.size() - 1];
-                motorData.pop_back();
-                motorData.pop_back();
-
-                publishMotor(_Cmd::CMD_VOLTAGE, motorData);
-                publishBattery(_Cmd::CMD_VOLTAGE, batteryData);
-
-                break;
-            case _Cmd::CMD_CURRENT:
-
-                if (convertBytesToFloat(data, motorData, nb_thruster + nb_battery) < 0)
-                {
-                    std::cerr << "ERROR in the message. Dropping CURRENT packet" << std::endl;
-                    return;
-                }
-
-                batteryData[0] = motorData[motorData.size() - 2];
-                batteryData[1] = motorData[motorData.size() - 1];
-                motorData.pop_back();
-                motorData.pop_back();
-
-                publishMotor(_Cmd::CMD_CURRENT, motorData);
-                publishBattery(_Cmd::CMD_CURRENT, batteryData);
-                break;
-            case _Cmd::CMD_TEMPERATURE:
-
-                if (convertBytesToFloat(data, motorData, nb_thruster + nb_battery) < 0)
-                {
-                    std::cerr << "ERROR in the message. Dropping TEMPERATURE packet" << std::endl;
-                    return;
-                }
-
-                batteryData[0] = motorData[motorData.size() - 2];
-                batteryData[1] = motorData[motorData.size() - 1];
-                motorData.pop_back();
-                motorData.pop_back();
-
-                publishMotor(_Cmd::CMD_TEMPERATURE, motorData);
-                publishBattery(_Cmd::CMD_TEMPERATURE, batteryData);
-                break;
-            case _Cmd::CMD_READ_MOTOR:                
-                publishMotorFeedback(data);
-                break;
-            default:
-                RCLCPP_WARN(this->get_logger(), "CMD Not identified");
-                break;
-        }
-    }
-
-    void RS485Provider::processAUV7PowerManagement(const uint8_t cmd, std::vector<uint8_t> (&psu_data)[4]){
-        std::vector<float> motorData;
-        motorData.reserve(8);
-        float batteryData[2];
-
-        if(checkNoEmptyVector(psu_data)){
-            switch(cmd){
-                case _Cmd::CMD_VOLTAGE:
-                {
-                    std::vector<float> convertData[4];
-                    for(size_t i=0; i<4; i++){    
-                        convertData[i].reserve(psu_data[i].size()/4);
-                        if (convertBytesToFloat(psu_data[i], convertData[i], psu_data[i].size()/4)<0)
-                        {
-                            std::cerr << "ERROR in the message. Dropping VOLTAGE packet" << std::endl;
-                            return;
-                        }
-                    }
-                    //8 motor data
-                    motorData.push_back(convertData[0].at(0));
-                    motorData.push_back(convertData[1].at(0));
-                    motorData.push_back(convertData[2].at(0));
-                    motorData.push_back(convertData[3].at(0));
-                    motorData.push_back(convertData[0].at(1));
-                    motorData.push_back(convertData[1].at(1));
-                    motorData.push_back(convertData[2].at(1));
-                    motorData.push_back(convertData[3].at(1));
-            
-                    //2 batteries data
-                    batteryData[0]=(convertData[0].at(3)+convertData[1].at(3))/2;
-                    batteryData[1]=(convertData[2].at(3)+convertData[3].at(3))/2;
-                    //publish voltages
-                    publishMotor(_Cmd::CMD_VOLTAGE, motorData);
-                    publishBattery(_Cmd::CMD_VOLTAGE, batteryData);
-                    break;
-                }
-                case _Cmd::CMD_CURRENT:
-                {
-                    std::vector<float> convertData[4];
-                    for(size_t i=0; i<4; i++){    
-                        convertData[i].reserve(psu_data[i].size()/4);
-                        if (convertBytesToFloat(psu_data[i], convertData[i], psu_data[i].size()/4)<0)
-                        {
-                            std::cerr << "ERROR in the message. Dropping CURRENT packet" << std::endl;
-                            return;
-                        }
-                        //std::cerr << "CONVERTDATA: "<<convertData[i].size() << std::endl;
-                    }
-                    //8 motor data
-                    motorData.push_back(convertData[0].at(0));
-                    motorData.push_back(convertData[1].at(0));
-                    motorData.push_back(convertData[2].at(0));
-                    motorData.push_back(convertData[3].at(0));
-                    motorData.push_back(convertData[0].at(1));
-                    motorData.push_back(convertData[1].at(1));
-                    motorData.push_back(convertData[2].at(1));
-                    motorData.push_back(convertData[3].at(1));
-            
-                    //2 batteries data
-                    batteryData[0]=(convertData[0].at(2)+convertData[1].at(2))/2;
-                    batteryData[1]=(convertData[2].at(2)+convertData[3].at(2))/2;
-                    //publish currents
-                    publishMotor(_Cmd::CMD_CURRENT, motorData);
-                    publishBattery(_Cmd::CMD_CURRENT, batteryData);
-                    break;
-                }
-                case _Cmd::CMD_READ_MOTOR:
-                {
-                    std::vector<uint8_t> motor_feedback;
-                    motor_feedback.push_back(psu_data[0].at(0));
-                    motor_feedback.push_back(psu_data[1].at(1));
-                    motor_feedback.push_back(psu_data[2].at(0));
-                    motor_feedback.push_back(psu_data[3].at(1));
-                    motor_feedback.push_back(psu_data[0].at(0));
-                    motor_feedback.push_back(psu_data[1].at(1));
-                    motor_feedback.push_back(psu_data[2].at(0));
-                    motor_feedback.push_back(psu_data[3].at(1));
-                    //motor feedback publish
-                    publishMotorFeedback(motor_feedback);
-                    break;
-                }
-                default:{
-                    RCLCPP_ERROR(this->get_logger(), "ERROR Unkown CMD for PWR management");
-                }
-            } 
-        } 
-    }
     void RS485Provider::readData()
     {
         // Delay for port opening
@@ -413,7 +167,8 @@ namespace rs485_port_manager
                 }
                 
                 _cvReaderParser.notify_all();
-            }std::this_thread::sleep_for(std::chrono::milliseconds(1));
+            }
+            std::this_thread::sleep_for(std::chrono::milliseconds(1));
         }
     }
 
@@ -421,7 +176,6 @@ namespace rs485_port_manager
     {
         
         std::unique_lock<std::mutex> _lockWriter(_mtxWriter);
-        time_t timestamp;
 
         // close the thread.
         while (_thread_control)
@@ -622,121 +376,5 @@ namespace rs485_port_manager
                 // packet dropped
             }
         }
-    }
-
-    bool RS485Provider::checkNoEmptyVector(std::vector<uint8_t> (&array)[4]){
-        for (const auto& v: array){
-            if(v.empty())return false;
-        }
-        return true;
-    }
-    int RS485Provider::convertBytesToFloat(const std::vector<uint8_t> &req, std::vector<float> &res, const size_t size)
-    {
-        uint8_t size_req = req.size();
-        if (size_req % 4 != 0) return -1;
-
-        _bytesToFloat converter;
-
-        for (uint8_t i = 0; i < size; ++i)  // shifting of 4 for each data
-        {
-            converter.bytes[0] = req[4 * i];
-            converter.bytes[1] = req[4 * i + 1];
-            converter.bytes[2] = req[4 * i + 2];
-            converter.bytes[3] = req[4 * i + 3];
-            res.push_back(converter.value);
-        }
-        return 0;
-    }
-    void RS485Provider::EnableDisableMotors(const std_msgs::msg::Bool &msg)
-    {
-        queueObject ser;
-        ser.cmd = _Cmd::CMD_ACT_MOTOR;
-        switch (ESC_SLAVE)
-        {
-            // AUV8 motor control
-            case _SlaveId::SLAVE_PWR_MANAGEMENT:
-                ser.slave = ESC_SLAVE;
-                ToggleMotors(msg.data, nb_thruster, ser.data);
-                _writerQueue.push_back(ser);
-                _cvReaderWriter.notify_all();
-                break;
-            // AUV7 motor control
-            case _SlaveId::SLAVE_ESC:
-                ser.slave = _SlaveId::SLAVE_PSU0;
-                ser.data.clear();
-                ToggleMotors(msg.data, nb_thruster / 4, ser.data);
-                _writerQueue.push_back(ser);
-
-                ser.slave = _SlaveId::SLAVE_PSU1;
-                ser.data.clear();
-                ToggleMotors(msg.data, nb_thruster / 4, ser.data);
-                _writerQueue.push_back(ser);
-
-                ser.slave = _SlaveId::SLAVE_PSU2;
-                ser.data.clear();
-                ToggleMotors(msg.data, nb_thruster / 4, ser.data);
-                _writerQueue.push_back(ser);
-
-                ser.slave = _SlaveId::SLAVE_PSU3;
-                ser.data.clear();
-                ToggleMotors(msg.data, nb_thruster / 4, ser.data);
-                _writerQueue.push_back(ser);
-                _cvReaderWriter.notify_all();
-                break;
-            default:
-                break;
-        }
-    }
-    void RS485Provider::ToggleMotors(const bool state, const uint8_t size, std::vector<uint8_t> &data)
-    {
-        if (state)
-        {
-            for (size_t i = 0; i < size; i++)
-            {
-                data.push_back(1);
-            }
-        }
-        else
-        {
-            for (size_t i = 0; i < size; i++)
-            {
-                data.push_back(0);
-            }
-        }
-    }
-    void RS485Provider::PwmCallback(const sonia_common_ros2::msg::MotorPwm &msg)
-    {
-        queueObject ser;
-        ser.cmd = _Cmd::CMD_PWM;
-        ser.slave = ESC_SLAVE;
-        ser.data.clear();
-        
-        ser.data.push_back(msg.motor1 >> 8);
-        ser.data.push_back(msg.motor1 & 0xFF);
-
-        ser.data.push_back(msg.motor2 >> 8);
-        ser.data.push_back(msg.motor2 & 0xFF);
-
-        ser.data.push_back(msg.motor3 >> 8);
-        ser.data.push_back(msg.motor3 & 0xFF);
-
-        ser.data.push_back(msg.motor4 >> 8);
-        ser.data.push_back(msg.motor4 & 0xFF);
-
-        ser.data.push_back(msg.motor5 >> 8);
-        ser.data.push_back(msg.motor5 & 0xFF);
-
-        ser.data.push_back(msg.motor6 >> 8);
-        ser.data.push_back(msg.motor6 & 0xFF);
-
-        ser.data.push_back(msg.motor7 >> 8);
-        ser.data.push_back(msg.motor7 & 0xFF);
-
-        ser.data.push_back(msg.motor8 >> 8);
-        ser.data.push_back(msg.motor8 & 0xFF);
-
-        _writerQueue.push_back(ser);
-        _cvReaderWriter.notify_all();
-        
     }
 }  // namespace rs485_port_manager
